refactor(SimuNEX): remove types namespace.
