+++ UnityProject/Assets/Scripts/Core/Derived/Faults/Stiction.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Diagnostics;
 
-namespace SimuNEX.Faults.Types
+namespace SimuNEX.Faults
 {
     /// <summary>
     /// A fault that applies a stiction effect to the value.
