+++ UnityProject/Assets/Scripts/Core/Interfaces/Fault.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 using UnityEngine;
 
 namespace SimuNEX.Faults
